Problem model update

+++ problem/models.py
@@ -18,8 +18,8 @@ class Problem(models.Model):
     problem_output = models.TextField()
     note = models.TextField(blank=True, null=True)
     difficulty = models.CharField(max_length=55, choices=DIFFICULTY_LEVEL, default='easy')
-    time_limit = models.IntegerField(default=1)
-    memory_limit = models.IntegerField(default=512)
+    time_limit = models.FloatField(default=1.0)
+    memory_limit = models.FloatField(default=512)
     created_by = models.ForeignKey(User, on_delete=models.SET_NULL, null=True)
     created_at = models.DateTimeField(auto_now_add=True)
